Replace `Stringable|string` argument types with `string`-only
Methods that previously accepted `Stringable|string` as argument types
now only support `string`.

`Stringable` was added for convenience so that someone could do,
for example

```php
$user = $auth->getUser('uid');
$auth->updateUser($user, [...]);
```

While convenient, this adds overhead when processing these arguments.
For example, if a method expects a non-empty string, the SDK would
have to do a `trim((string) $arg)` and check if it's empty.

With this change, we can rely only on a `@var non-empty-string $arg`
docblock annotation.

```php
$user = $auth->getUser('uid');
$auth->updateUser($user->uid, [...]);
```